implementacion para borrar plantas

+++ l_DataAccess/planta.js
@@ -56,17 +56,35 @@ module.exports = {
         })
     },
 
-    deleteEntry : (id) => {
-        pool.getConnection((err, connection) => {
-            if(err) throw err
-            connection.query('DELETE FROM planta WHERE idPlanta = ?', id, (err, rows) => {
-                connection.release()
-                if (!err) {
-                    console.log(`planta con id ${id} ha sido eliminada`)
-                } else {
-                    console.log(err)
-                }
+    deleteEntry : (id, callback) => {
+        const dataPromise = new Promise((resolve, reject) => {
+            pool.getConnection((err, connection) => {
+                if(err) throw err
+                connection.query('select * from edificio where planta = ?', id, (err, rows) => {
+                    if (!err) {
+                        if(rows.length != 0){
+                            connection.release()
+                            resolve(false)
+                        }
+                        else {
+                            connection.query('DELETE FROM planta WHERE idPlanta = ?', id, (err, rows) => {
+                                connection.release()
+                                if (!err) {
+                                    console.log(`Planta con id ${id} ha sido eliminado`)
+                                    resolve(true)
+                                } else {
+                                    reject(console.log(err))
+                                }
+                            })
+                        }
+                    } else {
+                        reject(console.log(err))
+                    }
+                })
             })
+        })
+        dataPromise.then(isDeleted => {
+            callback(isDeleted)
         })
     }
 }
